_posixshmem: extract into module/_posixshmem/ following the _socket pattern
importing.rs shrinks by 71 lines.
