+++ pyre/pyre-interpreter/src/module/_posixshmem/mod.rs
@@ -1,0 +1,6 @@
+//! _posixshmem module — PyPy: pypy/module/_posixshmem/
+//!
+//! Backs `multiprocessing.shared_memory` on POSIX.
+
+pub mod interp_posixshmem;
+pub mod moduledef;
