+++ pyre/pyre-interpreter/src/module/mod.rs
@@ -8,6 +8,8 @@ pub mod _io;
 #[allow(non_snake_case)]
 pub mod _ctypes;
 #[allow(non_snake_case)]
+pub mod _posixshmem;
+#[allow(non_snake_case)]
 pub mod _socket;
 pub mod _sre;
 pub mod _weakref;
